+++ components/nfc/mfkey32/screen/src/main/java/com/flipperdevices/nfc/mfkey32/screen/viewmodel/MfKey32ViewModel.kt
@@ -161,7 +161,7 @@ class MfKey32ViewModel @VMInject constructor(
         val foundedKey = FoundedKey(
             nonce.sectorName,
             nonce.keyName,
-            "%012X".format(key)
+            key?.let { "%012X".format(it) }
         )
         existedKeysStorage.onNewKey(foundedKey)
     }