+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - [FIX] Decrease reconnect timeouts and lags detector timeout
 - [FIX] Device info mapping keys
 - [FIX] Wait disconnecting from device when restart rpc
+- [FIX] Replace radio type to sub number in device info
 - [REFACTOR] Migrate bottom bar to compose navigation
 - [REFACTOR] Bump Android Gradle Plugin
 - [REFACTOR] Fix detekt compose issues